requirements.txt
amqp==5.1.1
aniso8601==9.0.1
autopep8==2.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.2.3
Flask-RESTful==0.3.9
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
prompt-toolkit==3.0.38
pycodestyle==2.10.0
python-dotenv==1.0.0
pytz==2023.3
redis==3.5.3
six==1.16.0
tomli==2.0.1
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
